[FEAT]: 알람기능 상세페이지로 이동
