+++ app/src/main/java/com/example/hotplacecontactapp/AddContactDialogFragment.kt
@@ -81,11 +81,6 @@ class AddContactDialogFragment : DialogFragment() {
             addContactBtnNegative.setOnClickListener {
                 dismiss()
             }
-
-            val handler = Handler(Looper.getMainLooper())
-            addContactBtnNotification.setOnClickListener {
-                handler.postDelayed({ eventNotification() }, 3000)
-            }
         }
     }
 
@@ -141,42 +136,6 @@ class AddContactDialogFragment : DialogFragment() {
             listener?.onContactAdded()
             dismiss()
         }
-    }
-
-    fun eventNotification() {
-        val manager = activity?.getSystemService(NOTIFICATION_SERVICE) as NotificationManager
-        val builder: NotificationCompat.Builder
-
-        val channelId = "one-channel"
-        val channelName = "My Channel One"
-        val channel = NotificationChannel(channelId, channelName, NotificationManager.IMPORTANCE_DEFAULT).apply {
-            description = "My Channel One Description"
-            setShowBadge(true)
-            val uri: Uri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION)
-            val audioAttributes = AudioAttributes.Builder()
-                .setContentType(AudioAttributes.CONTENT_TYPE_SONIFICATION)
-                .setUsage(AudioAttributes.USAGE_ALARM)
-                .build()
-            setSound(uri, audioAttributes)
-            enableVibration(true)
-        }
-
-        manager.createNotificationChannel(channel)
-        builder = NotificationCompat.Builder(requireContext(), channelId)
-
-        val intent = Intent(requireContext(), MainActivity::class.java)
-        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-        val pendingIntent = PendingIntent.getActivity(requireContext(), 0, intent, PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_IMMUTABLE)
-
-        builder.run {
-            setSmallIcon(R.mipmap.ic_launcher)
-            setWhen(System.currentTimeMillis())
-            setContentTitle("연락처 알림")
-            setContentText("설정한 알림이 도착했습니다!!")
-            setContentIntent(pendingIntent)
-            setAutoCancel(true)
-        }
-        manager.notify(11, builder.build())
     }
 
 }
