+++ app/src/main/java/com/example/hotplacecontactapp/ContactDetailFragment.kt
@@ -1,12 +1,22 @@
 package com.example.hotplacecontactapp
 
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.PendingIntent
+import android.content.Context
+import android.content.Intent
+import android.media.AudioAttributes
+import android.media.RingtoneManager
 import android.net.Uri
 import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
 import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.core.app.NotificationCompat
 import androidx.core.content.res.ResourcesCompat
 import com.example.hotplacecontactapp.databinding.ActivityContactDetailFragmentBinding
 
@@ -71,14 +81,23 @@ class ContactDetailFragment : Fragment() {
         val instaId = binding.tvDetailInstarUri
 
 
-
         img.setImageURI(param2[0].profileImage)
         name.text = param2[0].name
         num.text = param2[0].phoneNumber
         instaId.text = param2[0].instaAddress
         address.text = param2[0].address
 
-
+        val handler = Handler(Looper.getMainLooper())
+        binding.addContactBtn5m.setOnClickListener {
+            handler.postDelayed({ eventNotification() }, 1000)
+        }
+        binding.addContactBtn30m.setOnClickListener {
+            handler.postDelayed({ eventNotification() }, 3000)
+        }
+        binding.addContactBtn1h.setOnClickListener {
+            handler.postDelayed({ eventNotification() }, 5000)
+        }
+        
     }
 
     companion object {
@@ -89,5 +108,41 @@ class ContactDetailFragment : Fragment() {
                     putParcelableArrayList(ARG_PARAM2, param2)
                 }
             }
+    }
+
+    fun eventNotification() {
+        val manager = activity?.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+        val builder: NotificationCompat.Builder
+
+        val channelId = "one-channel"
+        val channelName = "My Channel One"
+        val channel = NotificationChannel(channelId, channelName, NotificationManager.IMPORTANCE_DEFAULT).apply {
+            description = "My Channel One Description"
+            setShowBadge(true)
+            val uri: Uri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION)
+            val audioAttributes = AudioAttributes.Builder()
+                .setContentType(AudioAttributes.CONTENT_TYPE_SONIFICATION)
+                .setUsage(AudioAttributes.USAGE_ALARM)
+                .build()
+            setSound(uri, audioAttributes)
+            enableVibration(true)
+        }
+
+        manager.createNotificationChannel(channel)
+        builder = NotificationCompat.Builder(requireContext(), channelId)
+
+        val intent = Intent(requireContext(), MainActivity::class.java)
+        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
+        val pendingIntent = PendingIntent.getActivity(requireContext(), 0, intent, PendingIntent.FLAG_CANCEL_CURRENT or PendingIntent.FLAG_IMMUTABLE)
+
+        builder.run {
+            setSmallIcon(R.mipmap.ic_launcher)
+            setWhen(System.currentTimeMillis())
+            setContentTitle("연락처 알림")
+            setContentText("설정한 알림이 도착했습니다!!")
+            setContentIntent(pendingIntent)
+            setAutoCancel(true)
+        }
+        manager.notify(11, builder.build())
     }
 }
