+++ app/src/main/res/layout/addcontact_layout_dialog.xml
@@ -105,49 +105,7 @@
         android:paddingStart="4dp"
         android:textColor="@android:color/holo_red_light"
         android:textSize="14sp" />
-    <Button
-        android:id="@+id/addContact_btn_notification"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="알람확인버튼"/>
 
-<!--    <com.google.android.material.chip.ChipGroup-->
-<!--        android:id="@+id/addContact_chipGroup"-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        app:singleSelection="true">-->
-
-<!--        <com.google.android.material.chip.Chip-->
-<!--            android:id="@+id/addContact_chip_off"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_marginEnd="4dp"-->
-<!--            style="@style/Widget.MaterialComponents.Chip.Choice"-->
-<!--            android:text="알림 OFF" />-->
-<!--        <com.google.android.material.chip.Chip-->
-<!--            android:id="@+id/addContact_chip_5m"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_marginEnd="4dp"-->
-<!--            style="@style/Widget.MaterialComponents.Chip.Choice"-->
-<!--            android:text="5분 후 알람" />-->
-
-<!--        <com.google.android.material.chip.Chip-->
-<!--            android:id="@+id/addContact_chip_30m"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_marginEnd="4dp"-->
-<!--            style="@style/Widget.MaterialComponents.Chip.Choice"-->
-<!--            android:text="30분 후 알람" />-->
-
-<!--        <com.google.android.material.chip.Chip-->
-<!--            android:id="@+id/addContact_chip_1h"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_marginEnd="4dp"-->
-<!--            style="@style/Widget.MaterialComponents.Chip.Choice"-->
-<!--            android:text="1시간 후 알람" />-->
-<!--    </com.google.android.material.chip.ChipGroup>-->
 
     <LinearLayout
         android:layout_width="wrap_content"
